requirements.txt
streamlit
openai
pandas
numpy
matplotlib
seaborn
langchain
python-dotenv
scikit-learn
plotly